core-compat-api: spit out explicit legacy extension converter
